Formatting + Week 1 done

+++ Minigin/Font.cpp
@@ -1,13 +1,13 @@
-#include <stdexcept>
-#include <SDL_ttf.h>
 #include "Font.h"
+#include <SDL_ttf.h>
+#include <stdexcept>
 
 TTF_Font *dae::Font::GetFont() const {
     return m_font;
 }
 
-dae::Font::Font(const std::string &fullPath, unsigned int size) : m_font(nullptr) {
-    m_font = TTF_OpenFont(fullPath.c_str(), size);
+dae::Font::Font(const std::string& fullPath, unsigned int size) : m_font(nullptr) {
+    m_font = TTF_OpenFont(fullPath.c_str(), static_cast<int>(size));
     if (m_font == nullptr) {
         throw std::runtime_error(std::string("Failed to load font: ") + SDL_GetError());
     }
